Ignore axis jitter.
Some gamepad axes produce a constant stream of events with little to no
difference in value, clogging the event queue and making the library extremely
slow, even making it lose important events. The commit fixes the problem by
making the library ignore any changes less than the maximum allowed jitter,
which is set to 1% of the value's dynamic range.