linux: More correct axis value normalization
The previous implementation relied on the assumption that either A) min is 0, or B) the
midpoint is 0. For devices that don't conform to that assumption, surprising things happen.
For instance, consider the behavior using a (min, max) of (20, 235). A value of 127, which
is expected to be neutral (~0.0), is normalized as 0.56. That's not good.

While I don't have first-hand experience with issues caused by assumption B, it's
nonetheless another avenue for surprises. If min is negative, 0 is always mapped to 0,
irrespective of the actual range.

There's an additional assumption made that the axis value will actually be within its
reported range. According to section 1.3 of the kernel docs for input programming
(https://www.kernel.org/doc/Documentation/input/input-programming.txt), min/max are values
"which the joystick *must* be able to reach, no problem if it sometimes reports more, but it
must be able to always reach the min and max values". A divide by zero can occur if min is 0
and the axis value is lower than 0, and it's entirely possible to produce values falling
outside the desired (-1.0, 1.0) range.

This revised implementation works for any range, so long as max > min, which I'd say is a much
safer assumption (and at that point, we could blame the driver author). Values are also
clamped to (-1.0, 1.0), ensuring application developers don't receive anything they aren't
prepared for. As an added bonus, IMO the new math is easier to follow.