Replace NativeEvCode with EvCode
Old NativeEvCode was a type alias to u16 and had a lot of problems.
New EvCode guarantee to refer only to one gamepad element (with
NativeEvCode, 15 could be either button or axis) and is strongly
typed. This change required a lot of refactoring, including changes to
public API.

This commit also change API for Mapping, which no longer overload
IndexMut operator. Mapping example was removed, because it required
updating to new Mapping API and I plan to bring it back improved later,
when we will support mappings form axis events to button events.

Test which checked all mappings from database was removed because it
relied too much on NativeEvCode begin u16 to fix it quickly.