Limit udev scan to the input subsystem
Fixes #111. This mirrors a similar fix in RetroArch
(https://github.com/libretro/RetroArch/pull/12728) for a similar issue
(https://github.com/libretro/RetroArch/issues/10408).

The crux of the problem is that we attempt to scan a node that
represents the battery of the controller, but the scan hangs due to a
firmware issue. As interacting with the battery is not necessary at this
stage, we can resolve the issue by limiting ourselves to the input
subsystem.