Release D-pad button if the direction of the axis changes
Gamepads may send a change on the axis direction (from -1.0 to 1.0 or
vice-versa) without sending an event for the neutral direction (0.0) for
this axis.
We currently don't handle this behavior, which is why there might be 2
opposite D-pad buttons pressed at the same time, because the internal
gamepad state doesn't release a button.

This commit fixes the behavior by releasing the specific buttons when
the direction of the axis changes.