Fix panic on MacOS
Ensure the device can be opened early in device_matching_cb() to filter it from the vec of DeviceInfo which is used to derive the IDs.

This fixes an issue when a device (3DConnexion Mouse) with kHIDPage_VendorDefined page is connected it causes open() to fail resulting in ID mismatches.