core: Expand `windows` version range to include `0.60` release
See the latest windows-rs crate releases at
https://github.com/microsoft/windows-rs/releases/tag/0.62.0 (tag has the
wrong number).

The argument type for the closure in `EventHandler` changed from
`&Option<T>` to `Ref<T>` which is basically a "borrow of" an `Option`
internally.  It exposes a `.as_ref()` returning `&T`, meaning the new
API is still code-compatible with older `windows` releases even if the
type of `g` is different.